api: reassembly: add reassembly offload
Add support for inline fragmentation reassembly offload in PKTIN
and IPSEC.

When the offload is enabled in pktin, reassembly would be attempted for
packets received on the pktio interface. When the offload is enabled in
IPSEC, reassembly would be attempted after IPSEC processing of inbound
packets. Support for this offload can be verified using corresponding
capability.

Application would be allowed to configure the below options,
1. Enable/disable the offload
2. Maximum wait time for reassembly
3. Maximum number of fragments

In case of success, the reassembled packet would be a regular ODP packet
thereby eliminating the need to have reassembly in the application.

For incomplete reassembly case, the partial reassembled fragments need
to be retrieved. An odp_packet_t with status as
'ODP_REASS_PACKET_INCOMPLETE' can be passed to odp_reass_result() for
getting the partial reassembled fragments. In addition to retrieving
fragments, the API returns additional information from the reassembly
attempt, including the time spent by the fragments in IPSEC/PKTIO
waiting for reassembly. This allows application to track reassembly
timeouts more accurately, when attempting reassembly on such packets.

Signed-off-by: Anoob Joseph <anoobj@marvell.com>