api: pktio: lso: add support for exact segment sizes
Add an LSO profile flag (and a related capability flag) to request
that all non-last segments created by the custom protocol LSO have
exactly the maximum payload length and may not be shorter.

This makes it possible for an application to control the exact
segment sizes for custom protocols that may require it.

Signed-off-by: Janne Peltonen <janne.peltonen@nokia.com>